Add kotlin example and more RoboLink Implementation Add logging and jblas dependencies to kotlin project Mod python example to work with standard input Add python gitignore, add to kotlin gitignore Mat -> DoubleMatrix Correct bug in NumberedEnum classes

+++ Kotlin/src/main/kotlin/com/robodk/api/Conversions.kt
@@ -29,8 +29,8 @@ fun ByteArray.toInt(): Int {
         throw IllegalArgumentException("Need 8 bytes to convert an Int")
     }
 
-    return (this[0].toLong() shl 24) +
-            (this[1].toLong() shl 16) +
-            (this[2].toLong() shl 8) +
-            this[3].toLong()
+    return (this[0].toInt() shl 24) +
+            (this[1].toInt() shl 16) +
+            (this[2].toInt() shl 8) +
+            this[3].toInt()
 }
